bringing back old palette
