+++ tools/snapshot-admin/palette.css
@@ -1,109 +1,48 @@
-/* Palette Styles - Aligned with Project Design System */
 body {
-  font-family: var(--body-font-family);
-  color: var(--color-text);
-  background-color: var(--color-background);
-  padding: var(--spacing-200);
-  margin: 0;
-  font-size: var(--body-size-s);
-  line-height: var(--line-height-s);
-}
-
-body * {
-  box-sizing: border-box;
+  font-family: Arial, sans-serif;
+  color: #DBDBDB;
+  padding: 0 8px;
 }
 
 p, details {
-  margin: var(--spacing-100) 0;
-  font-size: var(--body-size-s);
-  font-family: var(--body-font-family);
+  margin: 4px 0;
+  font-size: 12px;
+  font-family: adobe-clean, "Adobe Clean", "Source Sans Pro", -apple-system, system-ui, sans-serif;
 }
 
 details {
   display: inline-block;
-  margin-top: var(--spacing-200);
+  margin-top: 10px;
 }
 
 button {
-  color: var(--color-text);
-  background-color: var(--color-background);
-  border-radius: var(--rounding-s);
-  border: var(--border-s) solid var(--gray-300);
-  padding: var(--spacing-100) var(--spacing-200);
-  font-size: var(--body-size-s);
-  font-family: var(--body-font-family);
-  cursor: pointer;
-  transition: all 0.2s ease;
-}
-
-button:disabled {
-  opacity: 0.5;
-  cursor: not-allowed;
-}
-
-button:hover:not(:disabled) {
-  background-color: var(--gray-100);
-  border-color: var(--gray-400);
+  color: #dbdbdb;
+  background-color: unset;
+  border-radius: 16px;
+  border: 2px solid #DBDBDB;
+  padding: 4px 8px;
+  font-size: 12px;
 }
 
 .button-wrapper {
   display: inline-block;
   margin-left: 0;
-  margin-right: var(--spacing-200);
+  margin-right: 8px;
+}
+
+button:disabled {
+  display: none;
 }
 
 a {
-  color: var(--color-link);
-  text-decoration: none;
-}
-
-a:hover {
-  text-decoration: underline;
+  color: #dbdbdb;
 }
 
 p.buttons {
-  padding-top: var(--spacing-200);
-  height: auto;
-  display: flex;
-  gap: var(--spacing-100);
-  flex-wrap: wrap;
+padding-top: 8px;
+height: 32px;
 }
 
-/* Snapshot selector styles */
-#snapshot-selector-wrapper {
-  margin-bottom: var(--spacing-300);
-  padding-bottom: var(--spacing-200);
-  border-bottom: var(--border-s) solid var(--gray-300);
-}
-
-#snapshot-selector-wrapper label {
-  display: block;
-  margin-bottom: var(--spacing-100);
-  font-size: var(--body-size-s);
-  font-weight: var(--weight-bold);
-  color: var(--color-text);
-}
-
-#snapshot-select {
-  width: 100%;
-  padding: var(--spacing-100) var(--spacing-200);
-  border: var(--border-s) solid var(--gray-300);
-  border-radius: var(--rounding-s);
-  background-color: var(--color-background);
-  color: var(--color-text);
-  font-size: var(--body-size-s);
-  font-family: var(--body-font-family);
-}
-
-#snapshot-select:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-#snapshot-select option {
-  background-color: var(--color-background);
-  color: var(--color-text);
-}
 
 .loading {
   border: currentcolor 3px solid;
@@ -138,7 +77,7 @@ p.buttons {
   display: flex;
   justify-content: center;
   align-items: center;
-  background-color: var(--transparent-black-600);
+  background-color: #2228;
 }
 
 #spinner[aria-hidden="true"], #page-status-wrapper[aria-hidden="true"], #review-status-wrapper[aria-hidden="true"] {
@@ -149,36 +88,20 @@ details[open] {
   width: 100%;
 }
 
-details summary {
-  cursor: pointer;
-  font-weight: var(--weight-medium);
-  color: var(--color-text);
-}
-
-details summary:hover {
-  color: var(--color-link);
-}
-
 #page-list {
-  background-color: var(--transparent-black-200);
-  font-family: var(--code-font-family);
+  background-color: #0008;
+  font-family: 'Courier New', Courier, monospace;
   list-style: none;
-  padding: var(--spacing-100);
-  margin: var(--spacing-100) calc(-1 * var(--spacing-100));
-  border-radius: var(--rounding-s);
-  border: var(--border-s) solid var(--gray-300);
+  padding: 3px;
+  margin: 5px -5px;
+  border-radius: 5px;
+  border: 2px solid #DBDBDB;
   overflow-x: scroll;
 }
 
 #page-list li {
   display: flex;
   justify-content: space-between;
-  padding: var(--spacing-50) 0;
-  border-bottom: var(--border-s) solid var(--gray-200);
-}
-
-#page-list li:last-child {
-  border-bottom: none;
 }
 
 .snapshot-admin-oinw {
